Add a custom pre-commit config to be run by GitHub Actions.
We run pre-commit in GitHub Actions with a custom configuration that has 
the terraform-validate hook removed.  This is because terraform validate 
cannot currently run without accessing the remote state because of the 
way the providers are defined in terraform/providers.tf.  This is 
something that may be remedied in the future.  For more information, 
check out these two GitHub issues:
* hashicorp/terraform#15895
* hashicorp/terraform#15811